d5 part2
Reversed the order of the cargo stacks for this one (there's a lot of
context I'm not going to explain here)
Shift is cool and I appreciate that adding to the beginning of the array
can be referred to as "unshift". I'm sure there was a faster way to add
the individual elements of the shifted stack than to unshift and flatten
but it would've required more lines of code and the stacks are bounded
in size.